+++ SistemaBronx.Models/Producto.cs
@@ -10,9 +10,9 @@ public partial class Producto
 
     public int? IdCategoria { get; set; }
 
-    public int? PorcGanancia { get; set; }
+    public decimal? PorcGanancia { get; set; }
 
-    public int? PorcIva { get; set; }
+    public decimal? PorcIva { get; set; }
 
     public decimal CostoUnitario { get; set; }
